refactor: use typeorm utils to create a relation between customers and orders
