--- src/database/entity/index.ts
@@ -1,5 +1,0 @@
-// This is a barrel file
-
-export { Customer } from './Customer.js'
-export { Product } from './Product.js'
-export { Order } from './Order.js'
