+++ src/index.ts
@@ -1,7 +1,9 @@
 import { AppDataSource, connectToDatabase } from './database/connection.js'
-import { PremiumCustomer, RegularCustomer } from './database/entity/Customer.js'
+import { Customer, PremiumCustomer, RegularCustomer } from './database/entity/Customer.js'
+import { Order } from './database/entity/Order.js'
+import { Product } from './database/entity/Product.js'
 
-import { Customer, Order, Product } from './database/entity/index.js'
+
 import { fetchProducts } from './database/population.js'
 import { generateRandomQuantity } from './utils/generateRandomQuantity.js'
 
@@ -59,9 +61,9 @@ class Store {
 }
 
 await Store.establishConnection()
-const joe = new PremiumCustomer('John', 'joe@gmail.com', 1000)
-joe.modifyCart('addToCart', 84)
-joe.modifyCart('addToCart', 85)
+const joe = new PremiumCustomer('new', 'j4rrsda3ashe@gmail.com', 1000)
+await joe.modifyCart('addToCart', 84)
+await joe.modifyCart('addToCart', 85)
 await customerRepo.save(joe)
 await joe.modifyOrder('placeOrder')
 await joe.modifyOrder('payAllOrders')
